requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
google-generativeai>=0.8.0
pydantic>=2.0.0
